// src/features/leads-form/interfaces/LeadFormContent.ts
interface LeadFormContent {
  propertyType: string;
  purpose: string;
  minBudget: number;
  maxBudget: number;
  location: string;
  mustHaveItems: string;
  desiredItems: string;
  tenantName: string;
  tenantEmail: string;
  tenantPhone: string;
  tenantDocument: string;
  tenantPassword: string;
}

export default LeadFormContent;